move adapters to separated files refactored OnClickListener as Kotlin syntax require
